# Streamline Joni regexp functions usage of blocks and slices