@kaikalur 2021-08-18
Benchmark results:

New:
```
                sql_regexp_extract_alll ::  791.912 cpu ms ::    0B peak memory :: in   15K,      0B,   18.9K/s,      0B/s :: out     1,  10.7MB,       1/s,  13.5MB/s
                sql_regexp_replace ::  737.442 cpu ms ::    0B peak memory :: in   15K,      0B,   20.3K/s,      0B/s :: out     1,  21.7MB,       1/s,  29.4MB/s
```

Old:
```
                sql_regexp_extract_alll ::  971.138 cpu ms ::    0B peak memory :: in   15K,      0B,   15.4K/s,      0B/s :: out     1,  10.7MB,       1/s,    11MB/s
                 sql_regexp_replace :: 1180.955 cpu ms ::    0B peak memory :: in   15K,      0B,   12.7K/s,      0B/s :: out     1,  21.7MB,       0/s,  18.4MB/s
```