@kaikalur 2021-08-18
> Can we add a perf benchmark and the numbers? Thanks!

Done